refactor(taskrecord): clean up obsolete manual task type registry and tests
