+++ pkg/testutil/taskrecord/harness_test.go
@@ -218,110 +218,3 @@ func TestJobTestHarness_Profiling(t *testing.T) {
 		t.Errorf("expected Mem profile %s to exist", memPath)
 	}
 }
-
-func TestResolveTaskTypeFromTask(t *testing.T) {
-	testCases := []struct {
-		name     string
-		task     coretask.UntypedTask
-		wantType string
-		wantOk   bool
-	}{
-		{
-			name:     "nil task returns false",
-			task:     nil,
-			wantType: "",
-			wantOk:   false,
-		},
-		{
-			name: "task returning []*log.Log",
-			task: coretask.NewTask[[]*log.Log](
-				upstreamTaskID,
-				nil,
-				func(ctx context.Context) ([]*log.Log, error) { return nil, nil },
-			),
-			wantType: "[]*log.Log",
-			wantOk:   true,
-		},
-		{
-			name: "task returning []string",
-			task: coretask.NewTask[[]string](
-				downstreamTaskID,
-				nil,
-				func(ctx context.Context) ([]string, error) { return nil, nil },
-			),
-			wantType: "[]string",
-			wantOk:   true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			gotType, gotOk := ResolveTaskTypeFromTask(tc.task)
-			if diff := cmp.Diff(tc.wantOk, gotOk); diff != "" {
-				t.Errorf("ResolveTaskTypeFromTask() ok mismatch (-want +got):\n%s", diff)
-			}
-			if tc.wantOk {
-				if diff := cmp.Diff(tc.wantType, gotType.String()); diff != "" {
-					t.Errorf("ResolveTaskTypeFromTask() type string mismatch (-want +got):\n%s", diff)
-				}
-			}
-		})
-	}
-}
-
-func TestResolveTaskTypeFromTaskSet(t *testing.T) {
-	task1 := coretask.NewTask[[]*log.Log](
-		upstreamTaskID,
-		nil,
-		func(ctx context.Context) ([]*log.Log, error) { return nil, nil },
-	)
-	task2 := coretask.NewTask[[]string](
-		downstreamTaskID,
-		nil,
-		func(ctx context.Context) ([]string, error) { return nil, nil },
-	)
-	taskSet, err := coretask.NewTaskSet([]coretask.UntypedTask{task1, task2})
-	if err != nil {
-		t.Fatalf("failed to create task set: %v", err)
-	}
-
-	testCases := []struct {
-		name     string
-		taskRef  taskid.UntypedTaskReference
-		wantType string
-		wantOk   bool
-	}{
-		{
-			name:     "finds log slice task type",
-			taskRef:  upstreamTaskID.Ref(),
-			wantType: "[]*log.Log",
-			wantOk:   true,
-		},
-		{
-			name:     "finds string slice task type",
-			taskRef:  downstreamTaskID.Ref(),
-			wantType: "[]string",
-			wantOk:   true,
-		},
-		{
-			name:     "unknown task reference returns false",
-			taskRef:  taskid.NewTaskReference[int]("test/unknown"),
-			wantType: "",
-			wantOk:   false,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			gotType, gotOk := ResolveTaskTypeFromTaskSet(taskSet, tc.taskRef)
-			if diff := cmp.Diff(tc.wantOk, gotOk); diff != "" {
-				t.Errorf("ResolveTaskTypeFromTaskSet() ok mismatch (-want +got):\n%s", diff)
-			}
-			if tc.wantOk {
-				if diff := cmp.Diff(tc.wantType, gotType.String()); diff != "" {
-					t.Errorf("ResolveTaskTypeFromTaskSet() type string mismatch (-want +got):\n%s", diff)
-				}
-			}
-		})
-	}
-}
